tweak how we set the Agent context, to fix some weird state bugs
Occasionally, the enabled checkboxes for endpoints and the embedded server would lose their state and show disabled, even though the connection was still active. This was the case if you were running the agent for some time, but using other apps on the device.

I *suspect* that the MainActivity was being finalized, so when you re-entered Mercury it would call Agent.setContext() again and overwrite the endpoint/server parameter state in the Agent singleton.

Now, we will only overwrite those if they are null, which should set them when the app first launches then leave them alone. If the Agent is GC'd we have bigger problems anyway, because we'll lose track of the backing services.